fix(tui): keep subagent activity in the footer
